move logistics adapters to integration
