+++ mod/src/main/java/org/shsts/tinactory/AllCapabilities.java
@@ -14,8 +14,8 @@ import org.shsts.tinactory.api.machine.IProcessor;
 import org.shsts.tinactory.core.autocraft.integration.IPatternCellPort;
 import org.shsts.tinactory.core.logistics.IBytesProvider;
 import org.shsts.tinactory.core.logistics.IDigitalProvider;
-import org.shsts.tinactory.core.logistics.IFluidTanksHandler;
-import org.shsts.tinactory.core.logistics.IMenuItemHandler;
+import org.shsts.tinactory.integration.logistics.IFluidTanksHandler;
+import org.shsts.tinactory.integration.logistics.IMenuItemHandler;
 import org.shsts.tinactory.core.logistics.ISignalMachine;
 import org.shsts.tinactory.core.machine.ILayoutProvider;
 import org.shsts.tinycorelib.api.blockentity.IEventManager;
