+++ mod/src/main/java/org/shsts/tinactory/content/electric/BatteryBox.java
@@ -17,9 +17,9 @@ import org.shsts.tinactory.content.tool.BatteryItem;
 import org.shsts.tinactory.core.common.CapabilityProvider;
 import org.shsts.tinactory.core.electric.Voltage;
 import org.shsts.tinactory.core.gui.Layout;
-import org.shsts.tinactory.core.logistics.IMenuItemHandler;
+import org.shsts.tinactory.integration.logistics.IMenuItemHandler;
 import org.shsts.tinactory.core.logistics.StackHelper;
-import org.shsts.tinactory.core.logistics.WrapperItemHandler;
+import org.shsts.tinactory.integration.logistics.WrapperItemHandler;
 import org.shsts.tinactory.core.machine.ILayoutProvider;
 import org.shsts.tinactory.core.util.MathUtil;
 import org.shsts.tinycorelib.api.blockentity.IEventManager;
