+++ mod/src/main/java/org/shsts/tinactory/content/gui/TechMenu.java
@@ -13,7 +13,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.shsts.tinactory.content.gui.sync.OpenTechPacket;
 import org.shsts.tinactory.core.gui.Menu;
 import org.shsts.tinactory.core.logistics.StackHelper;
-import org.shsts.tinactory.core.logistics.WrapperItemHandler;
+import org.shsts.tinactory.integration.logistics.WrapperItemHandler;
 import org.shsts.tinycorelib.api.gui.MenuBase;
 
 import java.util.function.Consumer;
