+++ mod/src/main/java/org/shsts/tinactory/content/logistics/ElectricChest.java
@@ -16,9 +16,9 @@ import net.minecraftforge.items.IItemHandler;
 import net.minecraftforge.items.IItemHandlerModifiable;
 import org.shsts.tinactory.api.logistics.IPort;
 import org.shsts.tinactory.core.gui.Layout;
-import org.shsts.tinactory.core.logistics.ItemHandlerPort;
+import org.shsts.tinactory.integration.logistics.ItemHandlerPort;
 import org.shsts.tinactory.core.logistics.StackHelper;
-import org.shsts.tinactory.core.logistics.WrapperItemHandler;
+import org.shsts.tinactory.integration.logistics.WrapperItemHandler;
 import org.shsts.tinactory.core.util.MathUtil;
 import org.shsts.tinycorelib.api.blockentity.IEventManager;
 import org.shsts.tinycorelib.api.core.Transformer;
