+++ mod/src/main/java/org/shsts/tinactory/content/logistics/ElectricTank.java
@@ -14,10 +14,10 @@ import net.minecraftforge.common.util.LazyOptional;
 import net.minecraftforge.fluids.FluidStack;
 import net.minecraftforge.fluids.capability.IFluidHandler;
 import org.shsts.tinactory.core.gui.Layout;
-import org.shsts.tinactory.core.logistics.CombinedFluidTank;
-import org.shsts.tinactory.core.logistics.IFluidTanksHandler;
+import org.shsts.tinactory.integration.logistics.CombinedFluidTank;
+import org.shsts.tinactory.integration.logistics.IFluidTanksHandler;
 import org.shsts.tinactory.core.logistics.StackHelper;
-import org.shsts.tinactory.core.logistics.WrapperFluidTank;
+import org.shsts.tinactory.integration.logistics.WrapperFluidTank;
 import org.shsts.tinactory.core.util.MathUtil;
 import org.shsts.tinycorelib.api.core.Transformer;
 import org.shsts.tinycorelib.api.registrate.builder.IBlockEntityTypeBuilder;
