+++ mod/src/main/java/org/shsts/tinactory/content/logistics/MEDrive.java
@@ -23,9 +23,9 @@ import org.shsts.tinactory.core.autocraft.integration.NetworkPatternCell;
 import org.shsts.tinactory.core.common.CapabilityProvider;
 import org.shsts.tinactory.core.gui.Layout;
 import org.shsts.tinactory.core.logistics.IBytesProvider;
-import org.shsts.tinactory.core.logistics.IMenuItemHandler;
+import org.shsts.tinactory.integration.logistics.IMenuItemHandler;
 import org.shsts.tinactory.core.logistics.StackHelper;
-import org.shsts.tinactory.core.logistics.WrapperItemHandler;
+import org.shsts.tinactory.integration.logistics.WrapperItemHandler;
 import org.shsts.tinactory.core.machine.ILayoutProvider;
 import org.shsts.tinactory.core.machine.SimpleElectricConsumer;
 import org.shsts.tinactory.core.util.MathUtil;
