+++ mod/src/main/java/org/shsts/tinactory/content/machine/Workbench.java
@@ -22,9 +22,9 @@ import net.minecraftforge.items.wrapper.CombinedInvWrapper;
 import org.shsts.tinactory.AllTags;
 import org.shsts.tinactory.content.recipe.ToolRecipe;
 import org.shsts.tinactory.core.common.CapabilityProvider;
-import org.shsts.tinactory.core.logistics.IMenuItemHandler;
+import org.shsts.tinactory.integration.logistics.IMenuItemHandler;
 import org.shsts.tinactory.core.logistics.StackHelper;
-import org.shsts.tinactory.core.logistics.WrapperItemHandler;
+import org.shsts.tinactory.integration.logistics.WrapperItemHandler;
 import org.shsts.tinycorelib.api.blockentity.IEventManager;
 import org.shsts.tinycorelib.api.blockentity.IEventSubscriber;
 import org.shsts.tinycorelib.api.registrate.builder.IBlockEntityTypeBuilder;
