+++ mod/src/main/java/org/shsts/tinactory/content/multiblock/NuclearReactor.java
@@ -22,7 +22,7 @@ import org.shsts.tinactory.content.machine.Boiler;
 import org.shsts.tinactory.content.machine.IBoiler;
 import org.shsts.tinactory.content.tool.INuclearItem;
 import org.shsts.tinactory.core.logistics.StackHelper;
-import org.shsts.tinactory.core.logistics.WrapperItemHandler;
+import org.shsts.tinactory.integration.logistics.WrapperItemHandler;
 import org.shsts.tinactory.core.metrics.MetricsManager;
 import org.shsts.tinactory.core.multiblock.Multiblock;
 import org.shsts.tinactory.core.multiblock.MultiblockInterface;
