+++ mod/src/main/java/org/shsts/tinactory/core/gui/InventoryMenu.java
@@ -10,7 +10,7 @@ import net.minecraftforge.fluids.IFluidTank;
 import net.minecraftforge.fluids.capability.IFluidHandler;
 import net.minecraftforge.items.ItemHandlerHelper;
 import net.minecraftforge.items.wrapper.PlayerMainInvWrapper;
-import org.shsts.tinactory.core.logistics.IFluidTanksHandler;
+import org.shsts.tinactory.integration.logistics.IFluidTanksHandler;
 import org.shsts.tinactory.core.logistics.StackHelper;
 import org.shsts.tinycorelib.api.gui.MenuBase;
 import org.shsts.tinycorelib.api.network.IPacket;
