+++ mod/src/main/java/org/shsts/tinactory/integration/logistics/ItemHandlerPort.java
@@ -1,4 +1,4 @@
-package org.shsts.tinactory.core.logistics;
+package org.shsts.tinactory.integration.logistics;
 
 import com.mojang.logging.LogUtils;
 import javax.annotation.Nullable;
@@ -14,6 +14,7 @@ import org.shsts.tinactory.api.logistics.IPort;
 import org.shsts.tinactory.api.logistics.IPortFilter;
 import org.shsts.tinactory.api.logistics.IPortNotifier;
 import org.shsts.tinactory.api.logistics.PortType;
+import org.shsts.tinactory.core.logistics.StackHelper;
 import org.slf4j.Logger;
 
 import java.util.ArrayList;
