+++ mod/src/main/java/org/shsts/tinactory/integration/logistics/WrapperItemHandler.java
@@ -1,10 +1,11 @@
-package org.shsts.tinactory.core.logistics;
+package org.shsts.tinactory.integration.logistics;
 
 import javax.annotation.ParametersAreNonnullByDefault;
 import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.items.IItemHandlerModifiable;
 import net.minecraftforge.items.ItemStackHandler;
+import org.shsts.tinactory.core.logistics.PortNotifier;
 
 import java.util.Arrays;
 import java.util.function.Predicate;
